Move noise functions to noise-model
